+++ assets/js/customizer-liveview.js
@@ -6,18 +6,45 @@
  * Contains handlers to make Theme Customizer preview reload changes asynchronously.
  */
 
-( function( $ ) {
-    // Site title and description.
-    wp.customize( 'blogname', function( value ) {
+( function( $ , api ) {
+
+
+    // Site footer bg
+    wp.customize( 'onepress_footer_bg', function( value ) {
         value.bind( function( to ) {
-            $( '.site-title a' ).text( to );
+            $( '.site-footer' ).css( {
+                'background': to
+            } );
         } );
     } );
-    wp.customize( 'blogdescription', function( value ) {
+
+    // Site footer info bg
+    wp.customize( 'onepress_footer_info_bg', function( value ) {
         value.bind( function( to ) {
-            $( '.site-description' ).text( to );
+            $( '.site-footer .site-info, .site-footer .btt a' ).css( {
+                'background': to
+            } );
+
+            $( '.site-footer .site-info').css( {
+                color: 'rgba(255, 255, 255, 0.7)',
+            } );
+            $( '.site-footer .btt a, .site-footer .site-info a').css( {
+                color: 'rgba(255, 255, 255, 0.9)',
+            } );
         } );
     } );
+
+
+    /**
+     * Handle rendering of partials.
+     *
+     * @param {api.selectiveRefresh.Placement} placement
+     */
+    api.selectiveRefresh.bind( 'partial-content-rendered', function( placement ) {
+        $( window ).resize();
+    } );
+
+
     // Header text color.
     wp.customize( 'header_textcolor', function( value ) {
         value.bind( function( to ) {
@@ -35,4 +62,19 @@
             }
         } );
     } );
-} )( jQuery );
+
+
+    // Site footer widgets
+    wp.customize( 'onepress_btt_disable', function( value ) {
+        value.bind( function( to ) {
+            if ( to === true || to == 'true' ) {
+                $( '.site-footer .btt ' ).hide();
+            } else {
+                $( '.site-footer .btt ' ).show();
+            }
+        } );
+    } );
+    
+
+} )( jQuery , wp.customize );
+