+++ assets/js/customizer.js
@@ -305,6 +305,7 @@
 
 				var tpl = _tpl.replace( new RegExp(_id,"g"), id );
 				var template =  $( tpl );
+                template.find( 'textarea').removeAttr( 'rows').removeAttr( 'cols' );
 				$( "#"+id ).replaceWith( template );
 				// set content
 				$( '#'+id ).val( content );
@@ -353,6 +354,9 @@
 						}
 					}
 				};
+
+                //console.log( tmceInit );
+                tmceInit.plugins = tmceInit.plugins.replace('fullscreen,', '');
 
 				tinyMCEPreInit.mceInit[ id ] = tmceInit;
 
@@ -501,12 +505,26 @@
 				sync_id: control.editing_area,
 				init_instance_callback: function( editor ){
 					var w =  $( '#wp-'+control.editor_id+ '-wrap' );
-					$( '.wp-editor-tabs', w).append( '<button class="wp-switch-editor close-wp-editor"  type="button"><span class="dashicons dashicons-no-alt"></span></button>' );
-					w.on( 'click', '.close-wp-editor', function( e ) {
+                    $( '.wp-editor-tabs', w).append( '<button class="wp-switch-editor fullscreen-wp-editor"  type="button"><span class="dashicons"></span></button>' );
+                    $( '.wp-editor-tabs', w).append( '<button class="wp-switch-editor preview-wp-editor"  type="button"><span class="dashicons dashicons-visibility"></span></button>' );
+                    $( '.wp-editor-tabs', w).append( '<button class="wp-switch-editor close-wp-editor"  type="button"><span class="dashicons dashicons-no-alt"></span></button>' );
+                    w.on( 'click', '.close-wp-editor', function( e ) {
 						e.preventDefault();
 						control.editing_editor.removeClass( 'wpe-active' );
 						$( '.wp-js-editor-preview').removeClass( 'wpe-focus');
 					} );
+                    $( '.preview-wp-editor', w ).hover( function(){
+                        w.closest( '.modal-wp-js-editor').css( { opacity: 0 } );
+                    }, function(){
+                        w.closest( '.modal-wp-js-editor').css( { opacity: 1 } );
+                    } );
+                    w.on( 'click', '.fullscreen-wp-editor', function( e ) {
+                        e.preventDefault();
+                        w.closest( '.modal-wp-js-editor').toggleClass( 'fullscreen' );
+                        setTimeout( function(){
+                            $( window ).resize();
+                        }, 600 );
+                    } );
 				}
 			} );
 			control.editing_area.on( 'change', function() {
@@ -532,24 +550,16 @@
 
 		_resize: function(){
 			var control = this;
-			var w =  $( '#wp-'+control.editor_id+ '-wrap' );
-			var height = $( window ).height();
-			var tb_h = $( '.mce-toolbar-grp',  w).eq(0).height();
-			tb_h += $( '.wp-editor-tools', w ).eq(0).height();
-			tb_h += 80;
-			var width = $( window ).width();
+			var w =  $( '#wp-'+control.editor_id+ '-wrap');
+			var height = w.innerHeight();
+			var tb_h = w.find( '.mce-toolbar-grp' ).eq( 0 ).height();
+			tb_h += w.find( '.wp-editor-tools' ).eq( 0 ).height();
+            tb_h += 50;
+			//var width = $( window ).width();
 			var editor = tinymce.get( control.editor_id );
-			if ( width > 700 ) {
-				if ( width - 301 > 500 ) {
-					control.editing_editor.width( '' );
-				} else {
-					control.editing_editor.width( width - 330 );
-				}
-				editor.theme.resizeTo('100%', height - tb_h );
-			} else {
-				control.editing_editor.width( '' );
-				editor.theme.resizeTo('100%', height- ( tb_h + 150 ) );
-			}
+            control.editing_editor.width( '' );
+            editor.theme.resizeTo( '100%', height - tb_h );
+            w.find( 'textarea.wp-editor-area').height( height - tb_h  );
 		}
 
 	} );
@@ -723,12 +733,27 @@
 					sync_id: settings.editing_area,
 					init_instance_callback: function (editor) {
 						var w = $('#wp-' + settings.editor_id + '-wrap');
+                        $( '.wp-editor-tabs', w).append( '<button class="wp-switch-editor fullscreen-wp-editor"  type="button"><span class="dashicons"></span></button>' );
+                        $( '.wp-editor-tabs', w).append( '<button class="wp-switch-editor preview-wp-editor"  type="button"><span class="dashicons dashicons-visibility"></span></button>' );
 						$('.wp-editor-tabs', w).append('<button class="wp-switch-editor close-wp-editor"  type="button"><span class="dashicons dashicons-no-alt"></span></button>');
 						w.on('click', '.close-wp-editor', function (e) {
 							e.preventDefault();
 							settings.editing_editor.removeClass('wpe-active');
 							$( '.wp-js-editor-preview').removeClass( 'wpe-focus' );
 						});
+
+                        $( '.preview-wp-editor', w ).hover( function(){
+                            w.closest( '.modal-wp-js-editor').css( { opacity: 0 } );
+                        }, function(){
+                            w.closest( '.modal-wp-js-editor').css( { opacity: 1 } );
+                        } );
+                        w.on( 'click', '.fullscreen-wp-editor', function( e ) {
+                            e.preventDefault();
+                            w.closest( '.modal-wp-js-editor').toggleClass( 'fullscreen' );
+                            setTimeout( function(){
+                                $( window ).resize();
+                            }, 600 );
+                        } );
 					}
 				});
 				settings.editing_area.on('change', function () {
@@ -756,23 +781,14 @@
 			//----
 			settings._resize = function(){
 				var w =  $( '#wp-'+settings.editor_id+ '-wrap' );
-				var height = $( window ).height();
-				var tb_h = $( '.mce-toolbar-grp',  w).eq(0).height();
-				tb_h += $( '.wp-editor-tools', w ).eq(0).height();
-				tb_h += 80;
-				var width = $( window ).width();
+                var height = w.innerHeight();
+                var tb_h = w.find( '.mce-toolbar-grp' ).eq( 0 ).height();
+                tb_h += w.find( '.wp-editor-tools' ).eq( 0 ).height();
+                tb_h += 50;
 				var editor = tinymce.get( settings.editor_id );
-				if ( width > 700 ) {
-					if ( width - 301 > 500 ) {
-						settings.editing_editor.width( '' );
-					} else {
-						settings.editing_editor.width( width - 330 );
-					}
-					editor.theme.resizeTo('100%', height - tb_h );
-				} else {
-					settings.editing_editor.width( '' );
-					settings.theme.resizeTo('100%', height- ( tb_h + 150 ) );
-				}
+                settings.editing_editor.width( '' );
+                editor.theme.resizeTo( '100%', height - tb_h );
+                w.find( 'textarea.wp-editor-area').height( height - tb_h  );
 			};