+++ inc/customizer-controls.php
@@ -275,7 +275,7 @@ class OnePress_Editor_Custom_Control extends WP_Customize_Control
             <label>
                 <span class="customize-control-title"><?php echo esc_html( $this->label ); ?></span>
             </label>
-            <textarea class="wp-js-editor-textarea large-text" data-editor-mod="<?php echo esc_attr( $this->mod ); ?>" cols="20" rows="5" <?php $this->link(); ?>><?php echo esc_textarea( $this->value() ); ?></textarea>
+            <textarea class="wp-js-editor-textarea large-text" data-editor-mod="<?php echo esc_attr( $this->mod ); ?>" <?php $this->link(); ?>><?php echo esc_textarea( $this->value() ); ?></textarea>
             <p class="description"><?php echo $this->description ?></p>
         </div>
     <?php