+++ section-parts/section-features.php
@@ -3,13 +3,17 @@ $id       = get_theme_mod( 'onepress_features_id', esc_html__('features', 'onepr
 $disable  = get_theme_mod( 'onepress_features_disable' ) == 1 ? true : false;
 $title    = get_theme_mod( 'onepress_features_title', esc_html__('Features', 'onepress' ));
 $subtitle = get_theme_mod( 'onepress_features_subtitle', esc_html__('Why choose Us', 'onepress' ));
-
+if ( onepress_is_selective_refresh() ) {
+    $disable = false;
+}
 $data  = onepress_get_features_data();
 if ( !$disable && !empty( $data ) ) {
     $desc = get_theme_mod( 'onepress_features_desc' );
 ?>
+<?php if ( ! onepress_is_selective_refresh() ){ ?>
 <section id="<?php if ( $id != '') echo $id; ?>" <?php do_action('onepress_section_atts', 'features'); ?>
          class="<?php echo esc_attr(apply_filters('onepress_section_class', 'section-features section-padding section-meta onepage-section', 'features')); ?>">
+<?php } ?>
     <?php do_action('onepress_section_before_inner', 'features'); ?>
     <div class="container">
         <?php if ( $title ||  $subtitle || $desc ){ ?>
@@ -66,5 +70,8 @@ if ( !$disable && !empty( $data ) ) {
         </div>
     </div>
     <?php do_action('onepress_section_after_inner', 'features'); ?>
+
+<?php if ( ! onepress_is_selective_refresh() ){ ?>
 </section>
+<?php } ?>
 <?php } ?>