+++ section-parts/section-news.php
@@ -6,13 +6,18 @@ $onepress_news_subtitle  = get_theme_mod( 'onepress_news_subtitle', esc_html__('
 $onepress_news_number    = get_theme_mod( 'onepress_news_number', '3' );
 $onepress_news_more_link = get_theme_mod( 'onepress_news_more_link', '#' );
 $onepress_news_more_text = get_theme_mod( 'onepress_news_more_text', esc_html__('Read Our Blog', 'onepress' ));
+if ( onepress_is_selective_refresh() ) {
+    $onepress_news_disable = false;
+}
 ?>
 <?php if ( ! $onepress_news_disable  ) :
 
 $desc = get_theme_mod( 'onepress_news_desc' );
 ?>
+<?php if ( ! onepress_is_selective_refresh() ){ ?>
 <section id="<?php if ( $onepress_news_id != '' ) echo $onepress_news_id; ?>" <?php do_action( 'onepress_section_atts', 'news' ); ?> class="<?php echo esc_attr( apply_filters( 'onepress_section_class', 'section-news section-padding onepage-section', 'news' ) ); ?>">
-	<?php do_action( 'onepress_section_before_inner', 'news' ); ?>
+<?php } ?>
+    <?php do_action( 'onepress_section_before_inner', 'news' ); ?>
 	<div class="container">
 		<?php if ( $onepress_news_title ||  $onepress_news_subtitle ||  $desc ) { ?>
 		<div class="section-title-area">
@@ -30,7 +35,8 @@ $desc = get_theme_mod( 'onepress_news_desc' );
 						<?php
 						$query = new WP_Query(
 							array(
-								'posts_per_page' => $onepress_news_number
+								'posts_per_page' => $onepress_news_number,
+								'suppress_filters' => 0,
 							)
 						);
 						?>
@@ -65,7 +71,9 @@ $desc = get_theme_mod( 'onepress_news_desc' );
 		</div>
 	</div>
 	<?php do_action( 'onepress_section_after_inner', 'news' ); ?>
+<?php if ( ! onepress_is_selective_refresh() ){ ?>
 </section>
+<?php } ?>
 <?php endif;
 wp_reset_query();